+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix `Stats` field name from `network_stats` -> `networks`
 - Add missing clone implementations to some image api data types
 - All builder pattern methods now take an owned value and consume the builder on final build
+- Add a default implementation to `Isolation` to fix deserialization of `Info`
 
 # 0.6.0
 - `name` field of `ContainerCreateOpts` is now private. Use the `ContainerCreateOpts::builder` function that takes in a `name` parameter.